+++ CHANGELOG.md
@@ -1,20 +1,38 @@
 # Change Log
-All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org/).
+All notable changes to this project will be documented in this file. 
+This project adheres to [Semantic Versioning](http://semver.org/).
 
-## 0.1.3 — 2015-06-03
-### ADDED:
+## [1.0.0] — 2017-03-14
+### Added
+- Post API model
+- Category API model
+- Page API model
+- Media API model
+
+### Changed
+- Fix modman file
+- Move ACL configuration to adminhtml.xml
+- Refactor router
+- Move module to community code pool
+
+### Removed
+- Homepage support
+- Depricated post types support
+
+## [0.1.3] — 2015-06-03
+### Added
 - Missing sitemap model
 
-## 0.1.2 — 2015-06-01
-### ADDED:
+## [0.1.2] — 2015-06-01
+### Added
 - Posts and categories to sitemap
 
-## 0.1.1 — 2015-05-27
-### CHANGED:
+## [0.1.1] — 2015-05-27
+### Changed
 - Update readme file
 
-## 0.1.0 — 2015-05-27
-### ADDED:
+## [0.1.0] — 2015-05-27
+### Added
 - Change log file
 - Git ignore file
 - Module